feat: widen shared disassembly types for 32-bit CPUs
- Disassemble::disassemble takes addr: u32 and
  DisassembledInstruction::target_addr is Option<u32>; 16-bit CPU
  modules truncate to u16 at the top and keep their internal decode
  unchanged
- DisassembledInstruction::bytes grows [u8; 6] -> [u8; 10] (and the
  debug-UI fetch buffer with it) to fit the longest 68000 instruction,
  MOVE.L #imm,abs.l
- format_with_symbols resolves u32 targets and tries 8/6/4/2-digit hex
  forms widest-first when substituting
- DebugCpu impls pass the u32 address straight through (drops the
  per-CPU 'addr as u16' shims from the debug-surface widening)

+++ core/src/cpu/i8035/mod.rs
@@ -766,7 +766,7 @@ impl DebugCpu for I8035 {
     }
 
     fn debug_disassemble(&self, addr: u32, bytes: &[u8]) -> DisassembledInstruction {
-        <Self as crate::cpu::Disassemble>::disassemble(addr as u16, bytes)
+        <Self as crate::cpu::Disassemble>::disassemble(addr, bytes)
     }
 }
