+++ core/src/cpu/m6502/mod.rs
@@ -696,6 +696,6 @@ impl DebugCpu for M6502 {
     }
 
     fn debug_disassemble(&self, addr: u32, bytes: &[u8]) -> DisassembledInstruction {
-        <Self as crate::cpu::Disassemble>::disassemble(addr as u16, bytes)
+        <Self as crate::cpu::Disassemble>::disassemble(addr, bytes)
     }
 }
